+++ src/main/java/BotChat/BotChat.java
@@ -1,33 +1,48 @@
-package botChat;
+package BotChat;
 
 import java.util.Scanner;
 
-public class botChat {
+/**
+ * The main class for the botChat application, responsible for handling user input and managing tasks.
+ */
+public class BotChat {
+
     private Storage storage;
     private TaskList tasks;
     private Ui ui;
     private static final String FILE_PATH = "./botChat.txt";
 
-    public botChat(String filePath) {
+    /**
+     * Constructs a new botChat instance.
+     *
+     * @param filePath The file path for storing task data.
+     */
+    public BotChat(String filePath) {
         ui = new Ui();
         storage = new Storage(FILE_PATH);
     }
 
+    /**
+     * Runs the botChat application. Loads tasks, displays a welcome message, and handles user input.
+     */
     public void run() {
         tasks = storage.load();
         ui.showWelcomeMessage();
         userInput();
     }
 
+    /**
+     * Handles user input by continuously scanning and processing commands until the user exits.
+     */
     private void userInput() {
-        //Scanner to scan what the user is inputting
+        // Scanner to scan what the user is inputting
         Scanner scanner = new Scanner(System.in);
         try {
             while (true) {
                 String input = scanner.nextLine();
                 try {
                     handleInput(input);
-                } catch (botChatException e) {
+                } catch (BotChatException e) {
                     ui.showErrorMessage(e.getMessage());
                 }
                 Storage.saveTaskToHardDisk(tasks.getTasks());
@@ -37,8 +52,13 @@ public class botChat {
         }
     }
 
-    //Method to handle inputs
-    private void handleInput(String input) throws botChatException {
+    /**
+     * Handles different types of user input commands.
+     *
+     * @param input The user input command.
+     * @throws BotChatException If an error occurs during command processing.
+     */
+    private void handleInput(String input) throws BotChatException {
         Command command = Parser.getCommand(input.split(" ")[0]);
         switch (command) {
             case BYE:
@@ -56,22 +76,27 @@ public class botChat {
                 break;
             case EVENT:
                 tasks.addEventTask(input);
-                break;
+               break;
             case DEADLINE:
                 tasks.addDeadlineTask(input);
                 break;
             case TODO:
-                tasks.addTodoTask(input);
-                break;
+               tasks.addTodoTask(input);
+               break;
             case DELETE:
                 tasks.deleteTask(input);
                 break;
             case UNKNOWN:
-                throw new botChatException("Sorry, I do not understand that command. Please try again.");
+                throw new BotChatException("Sorry, I do not understand that command. Please try again.");
         }
     }
 
+    /**
+     * The main method to start the botChat application.
+     *
+     * @param args The command-line arguments (not used in this application).
+     */
     public static void main(String[] args) {
-        new botChat(FILE_PATH).run();
+        new BotChat(FILE_PATH).run();
     }
 }
