+++ src/main/java/BotChat/Parser.java
@@ -1,6 +1,16 @@
-package botChat;
+package BotChat;
 
+/**
+ * A utility class responsible for parsing user input and converting task information.
+ */
 public class Parser {
+
+    /**
+     * Parses a string representation of a command and returns the corresponding Command enum.
+     *
+     * @param input The string representation of the command.
+     * @return The Command enum associated with the input, or Command.UNKNOWN if not recognized.
+     */
     public static Command getCommand(String input) {
         try {
             return Command.valueOf(input.toUpperCase());
@@ -9,6 +19,12 @@ public class Parser {
         }
     }
 
+    /**
+     * Converts a string representation of a task into a Task object.
+     *
+     * @param line The string representation of a task.
+     * @return The Task object created based on the input string, or null if parsing fails.
+     */
     public static Task convertTask(String line) {
         String taskType = line.substring(1, 2);
         boolean isDone = line.substring(5, 6).equals("X");
