+++ src/main/java/BotChat/Task.java
@@ -1,0 +1,69 @@
+package BotChat;
+
+/**
+ * Represents a generic task in the botChat application.
+ */
+public class Task {
+
+    /**
+     * The description of the task.
+     */
+    protected String description;
+
+    /**
+     * Indicates whether the task is marked as done.
+     */
+    protected boolean isDone;
+
+    /**
+     * Constructs a Task object with the specified description.
+     *
+     * @param description The description of the task.
+     */
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    /**
+     * Gets the status icon indicating whether the task is done.
+     *
+     * @return The status icon ("X" if done, " " if not done).
+     */
+    public String getStatusIcon() {
+        return (isDone ? "X" : " ");
+    }
+
+    /**
+     * Gets the description of the task.
+     *
+     * @return The description of the task.
+     */
+    public String getDescription() {
+        return description;
+    }
+
+    /**
+     * Marks the task as completed.
+     */
+    public void mark() {
+        isDone = true;
+    }
+
+    /**
+     * Unmarks the task as completed.
+     */
+    public void unmark() {
+        isDone = false;
+    }
+
+    /**
+     * Provides a string representation of the task, including its status icon and description.
+     *
+     * @return A string representation of the task.
+     */
+    @Override
+    public String toString() {
+        return "[" + getStatusIcon() + "] " + description;
+    }
+}
