--- src/main/java/botChat/Command.java
@@ -1,13 +1,0 @@
-package botChat;
-
-public enum Command {
-    BYE,
-    LIST,
-    MARK,
-    UNMARK,
-    EVENT,
-    DEADLINE,
-    TODO,
-    DELETE,
-    UNKNOWN
-}
