--- src/main/java/botChat/Task.java
@@ -1,34 +1,0 @@
-package botChat;
-
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (isDone ? "X" : " ");
-    }
-    public String getDescription() {
-        return description;
-    }
-
-    //Method to mark completed tasks
-    public void mark() {
-        isDone = true;
-    }
-
-    //Method to unmark tasks
-    public void unmark() {
-        isDone = false;
-    }
-
-    //Method to print tasks
-    @Override
-    public String toString() {
-        return "[" + getStatusIcon() + "] " + description;
-    }
-}
