--- src/main/java/botChat/botChatException.java
@@ -1,7 +1,0 @@
-package botChat;
-
-public class botChatException extends Exception {
-    public botChatException(String message) {
-        super(message);
-    }
-}
