+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm botChat
+ Hello! I'm BotChat
  What can I do for you?
 ____________________________________________________________
